requirements.txt
flask
google-generativeai
python-dotenv
tenacity  # Add this for retries